requirements.txt
python-telegram-bot==13.11
python-dotenv==1.0.1
urllib3==1.26.15
redis==5.2.1
vk-api==11.9.9